changed to float

+++ script.js
@@ -45,41 +45,38 @@ document.getElementById('btn_0').addEventListener('click', function () {
 })
 
 document.getElementById('btn_sum').addEventListener('click', function () {
-    lastOperand = parseInt(inputWindow.value, 10);
+    lastOperand = parseFloat(inputWindow.value, 10);
     operation = 'sum';
     inputWindow.value = '';
 })
 
 document.getElementById('btn_div').addEventListener('click', function () {
-
-    lastOperand = parseInt(inputWindow.value, 10);
+    lastOperand = parseFloat(inputWindow.value, 10);
     operation = 'div';
 })
 
 document.getElementById('btn_div').addEventListener('click', function () {
-
-    lastOperand = parseInt(inputWindow.value, 10);
+    lastOperand = parseFloat(inputWindow.value, 10);
     operation = 'div';
     inputWindow.value = '';
 })
 
 document.getElementById('btn_mul').addEventListener('click', function () {
-
-    lastOperand = parseInt(inputWindow.value, 10);
+    lastOperand = parseFloat(inputWindow.value, 10);
     operation = 'mul';
     inputWindow.value = '';
 })
 
 document.getElementById('btn_sub').addEventListener('click', function () {
 
-    lastOperand = parseInt(inputWindow.value, 10);
+    lastOperand = parseFloat(inputWindow.value, 10);
     operation = 'sub';
     inputWindow.value = '';
 })
 
 document.getElementById('btn_calc').addEventListener('click', function () {
     if (operation === 'sum') {
-        const result = lastOperand + parseInt(inputWindow.value, 10);
+        const result = lastOperand + parseFloat(inputWindow.value, 10);
         operation = null;
         lastOperand = 0;
         inputWindow.value = result;
@@ -88,7 +85,7 @@ document.getElementById('btn_calc').addEventListener('click', function () {
         if (inputWindow.value == 0) {
             alert('На ноль нельзя делить');
         } else {
-            const result = lastOperand / parseInt(inputWindow.value, 10);
+            const result = lastOperand / parseFloat(inputWindow.value, 10);
             operation = null;
             lastOperand = 0;
             inputWindow.value = result;
@@ -96,13 +93,13 @@ document.getElementById('btn_calc').addEventListener('click', function () {
     }
 
     if (operation === 'mul') {
-        const result = lastOperand * parseInt(inputWindow.value, 10);
+        const result = lastOperand * parseFloat(inputWindow.value, 10);
         operation = null;
         lastOperand = 0;
         inputWindow.value = result;
     }
     if (operation === 'sub') {
-        const result = lastOperand - parseInt(inputWindow.value, 10);
+        const result = lastOperand - parseFloat(inputWindow.value, 10);
         operation = null;
         lastOperand = 0;
         inputWindow.value = result;
